# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SurahRevelationType =
  (typeof SurahRevelationType)[keyof typeof SurahRevelationType];

export const SurahRevelationType = {
  Meccan: "Meccan",
  Medinan: "Medinan",
} as const;

export interface Surah {
  id: number;
  number: number;
  /** Arabic name */
  name: string;
  /** English transliteration */
  nameSimple: string;
  /** English translation */
  nameTranslation: string;
  revelationType: SurahRevelationType;
  versesCount: number;
  pageNumber: number;
  juzNumber: number;
}

export interface Verse {
  id: string;
  /** e.g. "1:1" */
  verseKey: string;
  verseNumber: number;
  /** Arabic Quranic text */
  text: string;
  pageNumber: number;
  juzNumber: number;
  hizbNumber: number;
}

export interface VersesResponse {
  verses: Verse[];
  surah: Surah;
  totalCount: number;
}

export interface TafsirResponse {
  verseKey: string;
  /** Tafsir text */
  text: string;
  tafsirName: string;
  surahName: string;
  verseNumber: number;
  surahNumber: number;
}

export interface PageResponse {
  pageNumber: number;
  verses: Verse[];
  surahsOnPage: number[];
  /** Optional Mushaf page image URL */
  imageUrl?: string | null;
}

/**
 * Map of surah number to verse range string
 */
export type JuzVerseMapping = { [key: string]: unknown };

export interface Juz {
  juzNumber: number;
  /** Map of surah number to verse range string */
  verseMapping: JuzVerseMapping;
  firstVerseId: number;
  lastVerseId: number;
  versesCount: number;
}

export interface SearchResponse {
  surahs: Surah[];
  totalCount: number;
}

export interface Bookmark {
  id: number;
  pageNumber: number;
  surahNumber: number;
  verseNumber: number;
  surahName: string;
  note?: string | null;
  createdAt: string;
}

export interface CreateBookmarkBody {
  pageNumber: number;
  surahNumber: number;
  verseNumber: number;
  surahName: string;
  note?: string | null;
}

export interface ReadingProgress {
  id: number;
  lastPageNumber: number;
  lastSurahNumber: number;
  lastVerseNumber: number;
  lastSurahName: string;
  updatedAt: string;
}

export interface UpdateReadingProgressBody {
  lastPageNumber: number;
  lastSurahNumber: number;
  lastVerseNumber: number;
  lastSurahName: string;
}

export type AppSettingsTheme =
  (typeof AppSettingsTheme)[keyof typeof AppSettingsTheme];

export const AppSettingsTheme = {
  light: "light",
  dark: "dark",
} as const;

export interface AppSettings {
  id: number;
  theme: AppSettingsTheme;
  /** Font size percentage (80-150) */
  fontSize: number;
  showTafsir: boolean;
  defaultTafsirId: number;
  updatedAt: string;
}

export type UpdateSettingsBodyTheme =
  (typeof UpdateSettingsBodyTheme)[keyof typeof UpdateSettingsBodyTheme];

export const UpdateSettingsBodyTheme = {
  light: "light",
  dark: "dark",
} as const;

export interface UpdateSettingsBody {
  theme?: UpdateSettingsBodyTheme;
  fontSize?: number;
  showTafsir?: boolean;
  defaultTafsirId?: number;
}

export interface ReadingStats {
  totalBookmarks: number;
  lastReadPage: number;
  lastReadSurah: string;
  completedJuz: number;
  readingStreak: number;
  pagesRead: number;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export type GetSurahVersesParams = {
  page?: number;
  perPage?: number;
};

export type GetVerseTafsirParams = {
  /**
   * Tafsir ID (169 = Ibn Kathir Arabic)
   */
  tafsirId?: number;
};

export type SearchQuranParams = {
  q: string;
  language?: string;
};